feat(cli): move types to src and update corresponding paths
