+++ packages/cli/src/library/contants.ts
@@ -1,3 +1,3 @@
 export const API_BASE_URL =
   process.env.API_BASE_URL || 'https://hovcli.highoutput.io';
-export const API_TOKEN = process.env.HOVCLI_API_TOKEN || 'rBEqyGn9UBvmJ7hB';
+export const API_TOKEN = process.env.API_TOKEN || 'rBEqyGn9UBvmJ7hB';
