+++ packages/cli/src/library/render-database.ts
@@ -1,4 +1,4 @@
-import { Database, DatabaseStatus } from '../../types';
+import { Database, DatabaseStatus } from '../types'
 import R from 'ramda';
 import chalk from 'chalk';
