+++ packages/cli/src/library/render-deployments.ts
@@ -1,4 +1,4 @@
-import { Deployment, DeploymentStatus } from '../../types';
+import { Deployment, DeploymentStatus } from '../types'
 import R from 'ramda';
 import Str from '@supercharge/strings';
 import chalk from 'chalk';
