+++ packages/cli/src/modules/deployment/delete.ts
@@ -1,6 +1,6 @@
 import { Command } from 'commander';
 import renderDeployments from '../../library/render-deployments';
-import { Deployment, ObjectID, ResourceType } from '../../../types';
+import { Deployment, ObjectID, ResourceType } from '../../types'
 import * as client from '../../library/client';
 
 const command = new Command('delete');
