+++ packages/cli/src/modules/user/list.ts
@@ -1,6 +1,6 @@
 import { Command } from 'commander';
 import * as client from '../../library/client';
-import { ResourceType, User } from '../../../types';
+import { ResourceType, User } from '../../types'
 import renderUsers from '../../library/render-users';
 
 const command = new Command('list')
