+++ packages/cli/tsconfig.json
@@ -3,7 +3,6 @@
   "exclude": ["node_modules", "dist", "**/*.test.ts"],
   "include": [
     "src/**/*",
-    "types/*",
     "module.d.ts"
   ],
   "compilerOptions": {
